+++ crate/kmip/src/lib.rs
@@ -1,6 +1,8 @@
 #![allow(clippy::upper_case_acronyms)]
 //required to detect generic type in Serializer
 #![feature(min_specialization)]
+// To parse a slice
+#![feature(slice_take)]
 
 pub mod error;
 pub mod kmip;